Refactor home page to be a landing page and and it will take the user to the recipes, search functionality, update routing and URLs
